c5dec/assets/templates/project/pyproject.toml
[tool.poetry]
name = "c5dec"
version = "1.0"
description = "This new project is based on the C5-DEC method, templates, containerized environment and associated tools."
authors = ["Author <author@email.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
asciimatics = "^1.14.0"
python-i18n = "^0.3.9"
PyNaCl = "^1.5.0"
pyperclip = "^1.8.2"
pycryptodome = "^3.18.0"
python-docx = "^0.8.11"
doorstop = "3.0b9"
pandas = "^2.0.2"
blockdiag = "^3.0.0"
xlrd = "^2.0.1"
pylint = "^2.17.4"
graphviz = "^0.20.1"
pyyaml = "5.3.1"
flask = "3.0.3"

tabulate = "^0.9.0"
jupyter = "^1.1.1"
[tool.poetry.group.dev.dependencies]
pdoc3 = "^0.10.0"

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
